% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBvarsel-package.R
\docType{package}
\name{NBvarsel-package}
\alias{NBvarsel}
\alias{NBvarsel-package}
\title{NBvarsel: Variable Selection via Cross-Validated Net Benefit}
\description{
Performs exhaustive or groupwise (backward elimination) variable selection for binary outcome prediction models using cross-validated Net Benefit as the optimization criterion. Supports predictor costs, restricted cubic splines, interaction terms, permutation importance, and parallel computation. Includes visualizations for model comparison and variable importance.
}
\author{
\strong{Maintainer}: Lasai Barreñada \email{lasai.barrenadataleb@kuleuven.be} (\href{https://orcid.org/0000-0001-8020-0210}{ORCID})

}
\keyword{internal}
